refactor(downloads): satisfy output publication architecture gates
